Refactor: Centralize event save logic and rework event creation form
