+++ lib/core/core.dart
@@ -2,5 +2,5 @@
 // This provides a clean interface for importing core functionality
 
 export 'config/config_loader.dart';
-export '../data/local_data/data_loader.dart';
+export '../data/remote_data/load_data/data_loader.dart';
 export 'utils/date_utils.dart';
