+++ lib/core/di/dependency_injection.dart
@@ -1,7 +1,7 @@
 import 'package:get_it/get_it.dart';
 import 'package:sec/core/config/config_loader.dart';
 import 'package:sec/core/models/models.dart';
-import 'package:sec/data/local_data/data_loader.dart';
+import 'package:sec/data/remote_data/load_data/data_loader.dart';
 import 'package:sec/data/repositories/sec_repository_imp.dart';
 import 'package:sec/domain/repositories/sec_repository.dart';
 import 'package:sec/domain/use_cases/event_use_case.dart';
